Add new project step
